server/projects/analysis/requirements.txt
# Analysis Server Requirements
# server deploy
gunicorn==20.1.0

# for django
Django==3.1.12

#for django using mysql server
pymysql==1.0.2

# for celery
celery==4.3.0
vine==1.3.0
redis==3.3.11
flower==0.9.3

# for password
pyaes==1.6.1
pycryptodome==3.9.8
rsa==4.7.2

#for django permission
django-guardian==2.3.0

# for django template
Jinja2==2.11.3

# for rest api
djangorestframework==3.11.1
coreapi==2.3.3
Markdown==3.3.4
django-cors-headers==3.7.0
Pygments==2.8.1

# for document
drf-yasg==1.17.0

# for file upload and download
urllib3==1.26.7

django-filter==2.4.0

# for database import and export module
django-import-export==2.5.0

# for exception log
sentry-sdk==0.10.2

# for monitor
django-prometheus==2.0.0
